rendererGLES: use float value here,refs #1028

+++ src/rendererGLES/tr_init.c
@@ -37,7 +37,7 @@
 
 glconfig_t glConfig;
 qboolean   textureFilterAnisotropic = qfalse;
-int        maxAnisotropy            = 0;
+float        maxAnisotropy            = 0.0f;
 
 glstate_t glState;
